Jane

See merge request s03-webmobile1-sub2/s03p12b206!97

+++ WebPJT/frontend/src/page/post/PostListDetail.vue
@@ -23,7 +23,7 @@
                 <p
                   class="card-text"
                   style="font-size: 1rem; color: rgb(168, 168, 168); text-align: left; text-overflow:ellipsis;overflow: hidden;white-space: nowrap;"
-                >[{{post.location}}]{{post.sdate}}~{{post.edate}}</p>
+                >{{post.sdate}}~{{post.edate}}</p>
                   <a href="javascript:;" @click="test()" id="kakao-link-btn">  
                 <img src="//developers.kakao.com/assets/img/about/logos/kakaolink/kakaolink_btn_small.png" width="28px" />
                   </a>
@@ -55,14 +55,14 @@
                     <div class="d-flex justify-content-start">
                       <i
                           class="fas fa-heart select-button mr-2"
-                          style="text-align: right; font-size: 20px; color:red"
+                          style="text-align: right; font-size: 20px; color:crimson"
                         ></i> {{post.likecnt}}명이 좋아요를 눌렀습니다.
                     </div>
 
                 </div>
                 <div class="d-flex justify-content-end">
-                  <button type="button" class="btn btn-primary mr-1" data-toggle="modal" data-target="#BasketModal"><i class="fas fa-shopping-basket mr-2"></i>장바구니</button>
-                  <BasketModal />
+                  <button type="button" class="btn btn-primary mr-1" data-toggle="modal" data-target="#BasketModal" @click="alertbasket(post)"><i class="fas fa-shopping-basket mr-2"></i>장바구니</button>
+                  <!-- <BasketModal /> -->
                   <button class="btn btn-danger"><i class="far fa-hand-point-up mr-2"></i>바로구매</button>
                 </div>
               </div>
@@ -91,17 +91,14 @@
     <div data-spy="scroll" data-target="#navbar-example2" data-offset="0">
       <h4 id="item-info">상세 정보</h4>
       <p>{{post.detail}}</p>
+      <br>
       <h4 id="store-info">업체 정보</h4>
       <p>{{post.companyInfo}}</p>
+      <br>
       <h4 id="review">후기</h4>
-      <p>Ad leggings keytar, brunch id art party dolor labore. Pitchfork yr enim lo-fi before they sold out qui. Tumblr farm-to-table bicycle rights whatever. Anim keffiyeh carles cardigan. Velit seitan mcsweeney's photo booth 3 wolf moon irure. Cosby sweater lomo jean shorts, williamsburg hoodie minim qui you probably haven't heard of them et cardigan trust fund culpa biodiesel wes anderson aesthetic. Nihil tattooed accusamus, cred irony biodiesel keffiyeh artisan ullamco consequat.</p>
-      <p>Ad leggings keytar, brunch id art party dolor labore. Pitchfork yr enim lo-fi before they sold out qui. Tumblr farm-to-table bicycle rights whatever. Anim keffiyeh carles cardigan. Velit seitan mcsweeney's photo booth 3 wolf moon irure. Cosby sweater lomo jean shorts, williamsburg hoodie minim qui you probably haven't heard of them et cardigan trust fund culpa biodiesel wes anderson aesthetic. Nihil tattooed accusamus, cred irony biodiesel keffiyeh artisan ullamco consequat.</p>
-      <p>Ad leggings keytar, brunch id art party dolor labore. Pitchfork yr enim lo-fi before they sold out qui. Tumblr farm-to-table bicycle rights whatever. Anim keffiyeh carles cardigan. Velit seitan mcsweeney's photo booth 3 wolf moon irure. Cosby sweater lomo jean shorts, williamsburg hoodie minim qui you probably haven't heard of them et cardigan trust fund culpa biodiesel wes anderson aesthetic. Nihil tattooed accusamus, cred irony biodiesel keffiyeh artisan ullamco consequat.</p>
-      <p>Ad leggings keytar, brunch id art party dolor labore. Pitchfork yr enim lo-fi before they sold out qui. Tumblr farm-to-table bicycle rights whatever. Anim keffiyeh carles cardigan. Velit seitan mcsweeney's photo booth 3 wolf moon irure. Cosby sweater lomo jean shorts, williamsburg hoodie minim qui you probably haven't heard of them et cardigan trust fund culpa biodiesel wes anderson aesthetic. Nihil tattooed accusamus, cred irony biodiesel keffiyeh artisan ullamco consequat.</p>
-      
+     <br>
       <h4 id="qna">Q&A</h4>
-      <p>Ad leggings keytar, brunch id art party dolor labore. Pitchfork yr enim lo-fi before they sold out qui. Tumblr farm-to-table bicycle rights whatever. Anim keffiyeh carles cardigan. Velit seitan mcsweeney's photo booth 3 wolf moon irure. Cosby sweater lomo jean shorts, williamsburg hoodie minim qui you probably haven't heard of them et cardigan trust fund culpa biodiesel wes anderson aesthetic. Nihil tattooed accusamus, cred irony biodiesel keffiyeh artisan ullamco consequat.</p>
-    </div>
+      </div>
 
     <hr>
 
@@ -117,7 +114,7 @@
     
       
       <!-- 글 수정 삭제 -->
-      <div class="d-flex justify-content-end mt-3 mb-3">
+      <div class="d-flex justify-content-end mt-3 mb-3" v-if="this.$cookies.get('User') == this.post.email">
         <button class="btn btn-success" @click="goModify"><i class="far fa-edit mr-2"></i>수정하기</button>
         <button class="btn btn-danger" @click="goDelete"><i class="far fa-trash-alt mr-2"></i>삭제하기</button>
       </div>
@@ -129,10 +126,12 @@
 import axios from "axios";
 import '../../assets/css/postlistdetail.css'
 import PostUpdateVue from './PostUpdate.vue';
-import BasketModal from '../../components/modal/BasketModal.vue'
+// import BasketModal from '../../components/modal/BasketModal.vue'
 
 import CommentInput from '../../components/comment/CommentInput.vue'
 import CommentList from '../../components/comment/CommentList.vue'
+
+import Swal from 'sweetalert2'
 
 const baseURL = "http://localhost:8080";
 
@@ -140,7 +139,7 @@ export default {
   components: {
     CommentInput,
     CommentList,
-    BasketModal
+    // BasketModal
   },
   data(){
     return{
@@ -161,22 +160,22 @@ export default {
             container: '#kakao-link-btn',
             objectType: 'feed',
             content: {
-              title: '상세페이지 제목 호출',
-              description: '내용, 주로 해시태그',
-              imageUrl: document.images[0].src,
+              title: this.post.title, // 콘텐츠의 타이틀
+              description: this.post.activity,  // 콘텐츠 상세설명
+              imageUrl: document.images[0].src, // 썸네일 이미지
               link: {
                 webUrl: 'http://localhost:3000/#/posts/' + this.pid,
                 mobileWebUrl: 'https://developers.kakao.com'
               }
           },
           social: {
-            likeCount: 286,
-            commentCount: 45,
+            likeCount: 286, // LIKE 개수
+            commentCount: 45, // 댓글 개수
             sharedCount: 845
           },
           buttons: [
             {
-              title: 'Open!',
+              title: 'Open!',  // 버튼 제목
               link: {
                 mobileWebUrl: 'https://developers.kakao.com',
                 webUrl: 'http://localhost:3000/#/posts/' + this.pid     
@@ -205,19 +204,61 @@ export default {
       })
     },
     goDelete() {
-      axios.delete(`${baseURL}/post/delete/${this.$route.params.ID}`)
-        .then(() => {
-          alert('삭제 완료')
-          this.$router.push(`/posts`)
-        }).catch((error) => {
-          console.log(error.response.data)
-        })
+      Swal.fire({
+        width: 350,
+        text: '삭제하시겠습니까?',
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: '<a style="font-size:1rem; color:black">Delete</a>',
+        cancelButtonText: '<a style="font-size:1rem; color:black">Cancel</a>',
+      }).then((result) => {
+        if (result.value) {
+          const Toast = Swal.mixin({
+            toast: true,
+            position: 'top-end',
+            showConfirmButton: false,
+            timer: 3000,
+            timerProgressBar: true,
+            onOpen: (toast) => {
+              toast.addEventListener('mouseenter', Swal.stopTimer)
+              toast.addEventListener('mouseleave', Swal.resumeTimer)
+            }
+          })
+          Toast.fire({
+            icon: 'success',
+            title: '글이 삭제되었습니다.'
+          })
+          axios.delete(`${baseURL}/post/delete/${this.$route.params.ID}`)
+            .then(() => {
+              this.$router.push(`/posts`)
+            }).catch((error) => {
+              console.log(error.response.data)
+            })
+        }
+      })
     },
     createcomment(commentData) {
+      const Toast = Swal.mixin({
+        toast: true,
+        position: 'top-end',
+        showConfirmButton: false,
+        timer: 1500,
+        timerProgressBar: true,
+        onOpen: (toast) => {
+          toast.addEventListener('mouseenter', Swal.stopTimer)
+          toast.addEventListener('mouseleave', Swal.resumeTimer)
+        }
+      })
       axios.post(`${baseURL}/reply/register`,commentData)
         .then((response) => {
           commentData.content = ''
           this.fetchComment();
+          Toast.fire({
+            icon: 'success',
+            title: '댓글 작성 완료!'
+          })
         }).catch((error) => {
           console.log(error)
         })
@@ -231,12 +272,52 @@ export default {
         })
     },
     commentDelete(comment) {
-      axios.delete(`${baseURL}/reply/delete/${comment.rid}`)
-        .then((response) => {
-          this.fetchComment()
-        }).catch((error) => {
-          console.log(error.response.data)
-        })
+      Swal.fire({
+        width: 350,
+        text: "댓글을 삭제하시겠습니까?",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: '<a style="font-size:1rem; color:black">Delete</a>',
+        cancelButtonText: '<a style="font-size:1rem; color:black">Cancel</a>'
+      }).then((result) => {
+        if (result.value) {
+          const Toast = Swal.mixin({
+            toast: true,
+            position: 'top-end',
+            showConfirmButton: false,
+            timer: 1000,
+            timerProgressBar: true,
+            onOpen: (toast) => {
+              toast.addEventListener('mouseenter', Swal.stopTimer)
+              toast.addEventListener('mouseleave', Swal.resumeTimer)
+            }
+          })
+
+          Toast.fire({
+            icon: 'success',
+            title: '댓글이 삭제되었습니다.'
+          })
+          axios.delete(`${baseURL}/reply/delete/${comment.rid}`)
+            .then((response) => {
+              this.fetchComment()
+            }).catch((error) => {
+              console.log(error.response.data)
+            })
+        }
+      })
+    },
+    alertbasket(post){
+      Swal.fire({
+        title: `${post.title}`,
+        text: '장바구니에 담겼습니다.',
+        imageUrl: `${imgurl}`,
+        imageWidth: 400,
+        imageHeight: 200,
+        imageAlt: 'Custom image',
+      })
+      // alert(`'${title}'상품을 장바구니에 담았습니다!`)
     },
   },
